test(a2r): tt 档 27 件批量 bless(28 基线-真缺陷①已修)+tt 入复审清单——逐件 diff 评审留档(计划文件):mut 参数漂移6/括号化3/split物化6/as-u32字段2/算符优先级括号1/链式拆句1/File直写2/a2r_std前置2/fn指针clone1/doc注释1/as_str借用1;.cargo/config.toml tf 注记块增补复审清单须另跑 cargo tt (Plan 531)

+++ crates/auto-lang/test/cookbook/algorithms/003_sort_struct/sort_struct.expected.rs
@@ -8,7 +8,7 @@ struct Person {
 }
 
 fn main() {
-    let mut people = vec![Person { name: "Zoe".to_string(), age: 25 }, Person { name: "Al".to_string(), age: 60 }, Person { name: "John".to_string(), age: 1 }];
+    let mut people = vec![Person { name: "Zoe".to_string(), age: 25 as u32 }, Person { name: "Al".to_string(), age: 60 as u32 }, Person { name: "John".to_string(), age: 1 as u32 }];
